config dir 생성 및 폴더 정리
