+++ src/main/java/com/example/want/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.example.want.common;
+package com.example.want.config;
 
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
